requirements.txt
asgiref==3.8.1

attrs==25.4.

certifi==224.8.3

cffi==1.17.1

charset-normalizer==3.4.

colorama==.4.6

coreapi==2.3.3

coreschema==..4

cryptography==44..

defusedxml==.8.rc2

Django==5.1.4

django-cors-headers==4.6.

django-rest-framework==.1.

django-rest-swagger==2.2.

django-swagger-ui==.1.11

djangorestframework==3.15.2

djangorestframework-simplejwt==5.3.1

dotenv==.9.9

drf-spectacular==.28.

drf-yasg==1.21.8

idna==3.1

inflection==.5.1

itypes==1.2.

Jinja2==3.1.5

jsonschema==4.25.1

jsonschema-specifications==225.9.1

MarkupSafe==3..2

oauthlib==3.2.2

openapi-codec==1.3.2

packaging==24.2

pillow==11..

pycparser==2.22

PyJWT==2.1.1

python-dotenv==1.1.1

python-social-auth==.3.6

python3-openid==3.2.

pytz==224.2

PyYAML==6..2

qrcode==8.2

referencing==.37.

requests==2.32.3

requests-oauthlib==2..

rpds-py==.27.1

simplejson==3.19.3

social-auth-app-django==5.4.2

social-auth-core==4.5.4

sqlparse==.5.2

typing_extensions==4.12.2

tzdata==224.2

uritemplate==4.1.1

urllib3==2.2.3

whitenoise==6.8.2

